pyproject.toml
[tool.poetry]
name = "data-takehome-main"
version = "0.1.0"
description = "Milkmoovement Example Assignment"
authors = []
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
Flask = "^2.3.2"
requests = "^2.31.0"
boto3 = "^1.36.16"
snowflake-connector-python = "^3.13.2"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
